Modifique un par de boludeces y el archivo de listas ya no se maneja con offset sino con posiciones de los registros dentro del archivo

+++ trunk/src/CapaLogica/Indices/ArchivoListas.h
@@ -26,16 +26,16 @@ class ArchivoListas
         virtual int abrir(std::string directorioSalidaa, std::string fileNamee);
         //Abre el archivo con la direccion y el nombre que se le pasan por parametro
 
-        virtual long agregar(RegistroVariable *listaDeCanciones);
+        virtual unsigned short agregar(RegistroVariable *listaDeCanciones);
         //Agrega una lista al archivo y devuelve el offset de este
 
-        virtual int devolver(RegistroVariable *listaDeCanciones, long pos_lista);
+        virtual int devolver(RegistroVariable *listaDeCanciones, unsigned short pos_lista);
         //Guarda en la listaDeCanciones la lista que se encuentra en la posicion pos_lista dentro del archivo
 
         virtual int recontruir_listas(unsigned short* ref_listas, unsigned short cant_ref, int ID);
-        //Agrega el ID a las listas de las posiciones que guarda ref_listas
+        //Agrega el ID a las listas de las posiciones que guarda ref_listas. Se penso para los indices por Titulo y Autor
 
-        virtual int recontruir_listas(long ref_lista, RegistroVariable &listaModificada);
+        virtual int recontruir_listas(unsigned short ref_lista, RegistroVariable &listaModificada);
         //Modifico la lista que se pasa por referencia. Se penso para el indice invertido por frase
 
         virtual int eliminar(std::string directorioSalida, std::string fileNamee);
